fix: use i32 for make_interval hours param and test mark_account_dirty via HTTP
PostgreSQL's make_interval() expects int, not double precision. The
mark_account_dirty endpoint was binding hours_back as f64 which caused
a runtime error. Also updated the dirty monitor e2e test to call the
HTTP endpoint instead of using raw SQL to mark accounts dirty, and
fixed run_monitor_cycle call to match updated signature from main.

+++ nt-be/src/routes/monitored_accounts.rs
@@ -288,7 +288,7 @@ pub async fn mark_account_dirty(
         "#,
     )
     .bind(&account_id)
-    .bind(hours_back as f64)
+    .bind(hours_back as i32)
     .fetch_optional(&state.db_pool)
     .await
     .map_err(|e| {
